cleanup settings a bit
